cargo-psp: Remove Xargo.toml after xargo finishes
Previously, we attempted to wait on stdout, however this didn't work as
xargo prints to stderr. Then we attempted to wait on stderr without
changing the config to be `Stdio::piped`. This led to the current
behaviour of always panicking in the watcher thread. If we were to
correctly pipe xargo, we would no longer have a colored terminal output
as this is detected via TTY presence.

Instead of attempting to fix this, we revert to simply removing the
Xargo.toml file after the subprocess exits. Soon we will be able to move
away from xargo altogether and simply use the unstable `build-std` cargo
feature.